Moved tgui main function to main.c
